+++ lib/notifications/forum/__tests__/notificationRules.spec.ts
@@ -1,0 +1,10 @@
+import { generateUser } from 'testing/utils/users';
+
+describe('Notification Rules: isGroupEnabledForUser()', () => {
+  it('Returns true', () => {});
+  it('Returns false', async () => {
+    const user = await generateUser({
+      // notificationToggles: { rewards: false }
+    });
+  });
+});
