+++ lib/notifications/rewards/getRewardNotifications.ts
@@ -1,11 +1,11 @@
 import type { Page } from '@charmverse/core/prisma-client';
 import { prisma } from '@charmverse/core/prisma-client';
 
-import type { CardNotification } from './interfaces';
-import { notificationMetadataSelectStatement } from './utils';
+import type { BountyNotification } from '../interfaces';
+import { notificationMetadataSelectStatement } from '../utils';
 
-export async function getCardNotifications(userId: string): Promise<CardNotification[]> {
-  const cardNotifications = await prisma.cardNotification.findMany({
+export async function getBountyNotifications(userId: string): Promise<BountyNotification[]> {
+  const bountyNotifications = await prisma.bountyNotification.findMany({
     where: {
       notificationMetadata: {
         userId
@@ -14,12 +14,13 @@ export async function getCardNotifications(userId: string): Promise<CardNotifica
     select: {
       id: true,
       type: true,
-      personPropertyId: true,
-      card: {
-        include: {
+      applicationId: true,
+      bounty: {
+        select: {
+          status: true,
           page: {
             select: {
-              bountyId: true,
+              id: true,
               path: true,
               type: true,
               title: true
@@ -33,11 +34,12 @@ export async function getCardNotifications(userId: string): Promise<CardNotifica
     }
   });
 
-  return cardNotifications.map((notification) => {
+  return bountyNotifications.map((notification) => {
     const notificationMetadata = notification.notificationMetadata;
-    const page = notification.card.page as Page;
-    const cardNotification = {
+    const page = notification.bounty.page as Page;
+    const bountyNotification = {
       id: notification.id,
+      applicationId: notification.applicationId,
       createdAt: notificationMetadata.createdAt.toISOString(),
       createdBy: notificationMetadata.author,
       pageId: page.id,
@@ -46,15 +48,13 @@ export async function getCardNotifications(userId: string): Promise<CardNotifica
       spaceDomain: notificationMetadata.space.domain,
       spaceId: notificationMetadata.spaceId,
       spaceName: notificationMetadata.space.name,
-      pageType: page.type,
-      text: '',
+      status: notification.bounty.status,
       type: notification.type,
-      personPropertyId: notification.personPropertyId,
       archived: !!notificationMetadata.archivedAt,
       read: !!notificationMetadata.seenAt,
-      group: 'card'
-    } as CardNotification;
+      group: 'bounty'
+    } as BountyNotification;
 
-    return cardNotification;
+    return bountyNotification;
   });
 }
